special_robot/requirements.txt
nxt-python
pillow
requests
matplotlib
regex